making everything private that doesn't need to be public
